% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggallup.R
\name{ggallup}
\alias{ggallup}
\title{Apply Gallup-specific formatting to ggplot objects}
\usage{
ggallup(plot)
}
\arguments{
\item{plot}{A ggplot object to apply the formatting to}
}
\description{
Apply Gallup-specific formatting to ggplot objects
}
\examples{
\dontrun{ggallup(plot)}
}

