% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggallup_markdown.R
\name{ggallup_markdown}
\alias{ggallup_markdown}
\title{Apply Gallup-specific formatting to ggplot objects used in markdown documents}
\usage{
ggallup_markdown(plot)
}
\arguments{
\item{plot}{A ggplot object to apply the formatting to}
}
\description{
Apply Gallup-specific formatting to ggplot objects used in markdown documents
}
\examples{
\dontrun{ggallup(plot)}
}

